requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
yt-dlp>=2024.3.10